Rename Vessel/components/VesselSidebar/VesselReportingList to ReportingList
