+++ frontend/src/features/Vessel/components/VesselSidebar/ReportingList/DeletionConfirmationModal.tsx
