+++ frontend/src/features/Vessel/components/VesselSidebar/ReportingList/Summary/InfractionSuspicionSummary.tsx
