+++ frontend/src/features/Vessel/components/VesselSidebar/ReportingList/Summary/index.tsx
@@ -1,5 +1,5 @@
 import { Header, Zone } from '@features/Vessel/components/VesselSidebar/common_styles/common.style'
-import { InfractionSuspicionSummary } from '@features/Vessel/components/VesselSidebar/VesselReportingList/Summary/InfractionSuspicionSummary'
+import { InfractionSuspicionSummary } from '@features/Vessel/components/VesselSidebar/ReportingList/Summary/InfractionSuspicionSummary'
 import { customDayjs, Icon, THEME } from '@mtes-mct/monitor-ui'
 import styled from 'styled-components'
